(PC-23923) refactor(tutorial): move all onboarding pages in onboarding folder
