+++ src/features/tutorial/pages/AgeSelectionOther.tsx
@@ -7,7 +7,7 @@ import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
 import { AgeButton } from 'features/tutorial/components/AgeButton'
 import { useOnboardingContext } from 'features/tutorial/context/OnboardingWrapper'
-import { OnboardingPage } from 'features/tutorial/pages/OnboardingPage'
+import { OnboardingPage } from 'features/tutorial/pages/onboarding/OnboardingPage'
 import { NonEligible } from 'features/tutorial/types'
 import { analytics } from 'libs/analytics'
 import { storage } from 'libs/storage'
