+++ src/features/tutorial/pages/onboarding/OnboardingAgeInformation.native.test.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { navigate, reset } from '__mocks__/@react-navigation/native'
 import { OnboardingRootStackParamList } from 'features/navigation/RootNavigator/types'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
-import { AgeInformation } from 'features/tutorial/pages/AgeInformation'
+import { AgeInformation } from 'features/tutorial/pages/onboarding/OnboardingAgeInformation'
 import { CreditStatus } from 'features/tutorial/types'
 import { analytics } from 'libs/analytics'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
