+++ src/features/tutorial/pages/onboarding/OnboardingAgeInformation.web.test.tsx
@@ -2,7 +2,7 @@ import { StackScreenProps } from '@react-navigation/stack'
 import React from 'react'
 
 import { OnboardingRootStackParamList } from 'features/navigation/RootNavigator/types'
-import { AgeInformation } from 'features/tutorial/pages/AgeInformation'
+import { AgeInformation } from 'features/tutorial/pages/onboarding/OnboardingAgeInformation'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { render } from 'tests/utils/web'
