+++ src/features/tutorial/pages/onboarding/OnboardingWelcome.native.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { OnboardingWelcome } from 'features/tutorial/pages/OnboardingWelcome'
+import { OnboardingWelcome } from 'features/tutorial/pages/onboarding/OnboardingWelcome'
 import { analytics } from 'libs/analytics'
 import { storage } from 'libs/storage'
 import { fireEvent, render, waitFor } from 'tests/utils'
